Separated AttributableTests into 2 test classes.
