+++ Tests/AttributesTests/schema/AttributableTests.swift
@@ -57,9 +57,6 @@
 @testable import Attributes
 import XCTest
 
-// swiftlint:disable file_length
-// swiftlint:disable type_body_length
-
 /// Test class for ``Attributable`` default implementations.
 final class AttributableTests: XCTestCase {
 
@@ -265,276 +262,6 @@ final class AttributableTests: XCTestCase {
         XCTAssertNil(person.findProperty(path: path, in: person.data))
     }
 
-    /// Test whenChanged creates correct trigger.
-    func testWhenChangedPerformsCallback() throws {
-        let trigger = person.WhenChanged(SchemaAttribute(label: "first_name", type: .line)).makeUnavailable(
-            field: Field(name: "last_name", type: .line),
-            fields: Path(EmptyModifiable.self).attributes[0].attributes["person"].wrappedValue.complexFields
-        )
-        XCTAssertTrue(
-            try trigger.performTrigger(&person.data, for: AnyPath(Path(EmptyModifiable.self))).get()
-        )
-        guard let fields = person.data.attributes.first?.attributes["person"]?.complexFields else {
-            XCTFail("Cannot get fields")
-            return
-        }
-        XCTAssertEqual(
-            fields,
-            [
-                Field(name: "first_name", type: .line),
-                Field(name: "is_male", type: .bool)
-            ]
-        )
-    }
-
-    /// Test fields become available when a bool is true.
-    func testWhenTrueMakeAvailable() throws {
-        let trigger = person.WhenTrue(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeAvailable: SchemaAttribute(label: "age", type: .integer)
-        )
-        XCTAssertTrue(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        guard let fields = person.data.attributes.first?.attributes["person"]?.complexFields else {
-            XCTFail("Cannot get fields")
-            return
-        }
-        let expectedFields = [Field(name: "age", type: .integer)] + personFields
-        XCTAssertEqual(fields, expectedFields)
-    }
-
-    /// Checks whether the WhenTrue trigger works when a false value is present.
-    func testWhenTrueMakeAvailableWithFalseValue() throws {
-        let trigger = person.WhenTrue(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeAvailable: SchemaAttribute(label: "age", type: .integer)
-        )
-        guard var attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        attribute["is_male"] = .bool(false)
-        person.data.attributes[0].attributes["person"] = .complex(attribute, layout: personFields)
-        XCTAssertFalse(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexValue, attribute)
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexFields, personFields)
-    }
-
-    /// Test fields become available when a bool is false.
-    func testWhenFalseMakeAvailable() throws {
-        let trigger = person.WhenFalse(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeAvailable: SchemaAttribute(label: "age", type: .integer)
-        )
-        guard var attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        attribute["is_male"] = .bool(false)
-        person.data.attributes[0].attributes["person"] = .complex(attribute, layout: personFields)
-        XCTAssertTrue(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        guard let fields = person.data.attributes.first?.attributes["person"]?.complexFields else {
-            XCTFail("Cannot get fields")
-            return
-        }
-        let expectedFields = [Field(name: "age", type: .integer)] + personFields
-        XCTAssertEqual(fields, expectedFields)
-    }
-
-    /// Checks whether the WhenFalse trigger works when a true value is present.
-    func testWhenFalseMakeAvailableWithTrueValue() throws {
-        let trigger = person.WhenFalse(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeAvailable: SchemaAttribute(label: "age", type: .integer)
-        )
-        guard let attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        XCTAssertFalse(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexValue, attribute)
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexFields, personFields)
-    }
-
-    /// Test fields become unavailable when a bool is true.
-    func testWhenTrueMakeUnavailable() throws {
-        let trigger = person.WhenTrue(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeUnavailable: SchemaAttribute(label: "first_name", type: .line)
-        )
-        XCTAssertTrue(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        guard let fields = person.data.attributes.first?.attributes["person"]?.complexFields else {
-            XCTFail("Cannot get fields")
-            return
-        }
-        let expectedFields = [Field(name: "last_name", type: .line), Field(name: "is_male", type: .bool)]
-        XCTAssertEqual(fields, expectedFields)
-    }
-
-    /// Checks whether the WhenTrue trigger works when a false value is present.
-    func testWhenTrueMakeUnavailableWithFalseValue() throws {
-        let trigger = person.WhenTrue(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeUnavailable: SchemaAttribute(label: "first_name", type: .line)
-        )
-        guard var attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        attribute["is_male"] = .bool(false)
-        person.data.attributes[0].attributes["person"] = .complex(attribute, layout: personFields)
-        XCTAssertFalse(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexValue, attribute)
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexFields, personFields)
-    }
-
-    /// Test fields become unavailable when a bool is false.
-    func testWhenFalseMakeUnavailable() throws {
-        let trigger = person.WhenFalse(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeUnavailable: SchemaAttribute(label: "first_name", type: .line)
-        )
-        guard var attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        attribute["is_male"] = .bool(false)
-        person.data.attributes[0].attributes["person"] = .complex(attribute, layout: personFields)
-        XCTAssertTrue(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        guard let fields = person.data.attributes.first?.attributes["person"]?.complexFields else {
-            XCTFail("Cannot get fields")
-            return
-        }
-        let expectedFields = [Field(name: "last_name", type: .line), Field(name: "is_male", type: .bool)]
-        XCTAssertEqual(fields, expectedFields)
-    }
-
-    /// Checks whether the WhenFalse trigger works when a true value is present.
-    func testWhenFalseMakeUnavailableWithTrueValue() throws {
-        let trigger = person.WhenFalse(
-            SchemaAttribute(label: "is_male", type: .bool),
-            makeUnavailable: SchemaAttribute(label: "first_name", type: .line)
-        )
-        guard let attribute = person.data.attributes[0].attributes["person"]?.complexValue else {
-            XCTFail("Cannot get attribute.")
-            return
-        }
-        XCTAssertFalse(
-            try trigger.performTrigger(
-                &person.data,
-                for: AnyPath(
-                    Path(EmptyModifiable.self)
-                        .attributes[0]
-                        .attributes["person"]
-                        .wrappedValue
-                        .blockAttribute
-                        .complexValue["is_male"]
-                        .wrappedValue
-                )
-            )
-            .get()
-        )
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexValue, attribute)
-        XCTAssertEqual(person.data.attributes.first?.attributes["person"]?.complexFields, personFields)
-    }
-
     /// Test properties computed property works for all attributes.
     func testAllProperties() {
         let properties = SchemaProperties().properties
@@ -566,6 +293,3 @@ final class AttributableTests: XCTestCase {
     }
 
 }
-
-// swiftlint:enable type_body_length
-// swiftlint:enable file_length
